Make Exit() recycle task descriptors
- Puts task descriptor in FREE state, and adds it to the free queue.

- Increments the task descriptor's TID sequence number. The sequence
  numbers are all initialized to zero on kernel startup, and task
  creation uses the existing value. This way, the first task created
  with a given task descriptor gets sequence number zero, and the
  sequence number increases by one for each subsequent task that
  reuses that descriptor.

NB. the sequence number will overflow (and then TIDs will be
re-used) if a single task descriptor is used for more than 256
tasks. If it's REALLY important to never re-use TIDs, then we
can detect this situation and retire that task descriptor,
but I think it's likely this case won't happen in practice.