Rudimentary task timing.
Time for each task is printed on exit, optionally based on a setting
in struct kparam. This is on by default, but off in tests.

This prints the raw values from the 40-bit debug timer. Will want to
convert these to percentages later, but it's tricky to do it without
overflowing.