Implemented decrease-key for both priority queues.
This is desirable for Dijkstra's algorithm.

The implementation heavily restricts the values that can be stored in a
priority queue: if max size is N, only values 0 through N-1 can be
stored, and duplicates are not allowed. In order to store any other
data, an auxiliary array will be required.

The arguments to pqueue_add has been reordered to match the order of
arguments to pqueue_decreasekey.

Tests and the clock server have been updated to reflect the above.